backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
pydantic
mysql-connector-python
scikit-learn
pandas
joblib
langchain
chromadb
google-generativeai
huggingface-hub
transformers